Move descriptions into own file
