+++ v2/component.ts
@@ -1,7 +1,7 @@
 import type { Observable } from 'rxjs'
 import type { ButterfloatEvents, DefaultEvents } from '../events.js'
 import { Ring } from './ring.js'
-import { Mat, matType } from './mat.js'
+import { Mat } from './mat.js'
 
 /**
  * A Butterfloat Component provided properties and additional context-sensitive tools
@@ -125,157 +125,4 @@ export interface ButterfloatIntrinsicAttributes<
    * Immediately bind a boolean observable to the appearance of a class in classList.
    */
   immediateClassBind?: ClassBind
-}
-
-/*
-    Discussion: the "Description" types look a lot more verbose than a "standard" VNode
-    interface in a Virtual DOM. In this case, this is because this isn't intended for
-    virtual DOM usage. These should last only long enough to build a static DOM and bind
-    it, then tossed. There's no intention to have a diff/patch between trees of these.
-
-    On the other hand, these objects are still useful for DOM-less unit testing of
-    components.
-
-    So it makes sense to use full words. Users may work with these in their tests.
-*/
-
-/**
- * A Description that supports binding Children
- */
-export interface ChildrenBindDescription {
-  children: JsxChildren
-  childrenBind?: ChildrenBind
-  childrenBindMode?: ChildrenBindMode
-}
-
-/**
- * Description of a DOM element and its bindings
- */
-export interface ElementDescription<Bind = DefaultBind>
-  extends ChildrenBindDescription {
-  type: 'element'
-  element: string
-  attributes: Attributes
-  bind: Bind
-  immediateBind: Bind
-  events: DefaultEvents
-  styleBind: DefaultStyleBind
-  immediateStyleBind: DefaultStyleBind
-  classBind: ClassBind
-  immediateClassBind: ClassBind
-}
-
-/**
- * Description of a Component binding
- */
-export interface ComponentDescription extends ChildrenBindDescription {
-  type: 'component'
-  component: Component
-  properties: Attributes
-}
-
-/**
- * Description of a Fragment (the `<Fragment>` pseudo-component which powers `<></>` fragment notation)
- */
-export interface FragmentDescription {
-  type: 'fragment'
-  children: JsxChildren
-}
-
-/**
- * Description of the `<Children>` pseudo-component
- */
-export interface ChildrenDescription {
-  type: 'children'
-  context?: Mat<unknown>
-}
-
-/**
- * Description of the `<Static>` pseudo-component
- */
-export interface StaticDescription {
-  type: 'static'
-  element: Element
-}
-
-/**
- * Description of the `<Empty>` pseudo-component
- */
-export interface EmptyDescription {
-  type: 'empty'
-}
-
-/**
- * Description of the `<Comment>` pseudo-component
- */
-export interface CommentDescription {
-  type: 'comment'
-  comment: string
-}
-
-/**
- * A description of a node in a Butterfloat DOM tree
- */
-export type NodeDescription =
-  | ElementDescription
-  | ComponentDescription
-  | FragmentDescription
-  | ChildrenDescription
-  | StaticDescription
-  | EmptyDescription
-  | CommentDescription
-
-/**
- * A Component Context for Testing purposes
- */
-export interface TestComponentContext<Events = DefaultEvents> {
-  context: Mat<Events>
-  // Types here are just for examing test results
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  effects: Array<[Observable<unknown>, (item: any) => void]>
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  immediateEffects: Array<[Observable<unknown>, (item: any) => void]>
-}
-
-/**
- * Make a test context for testing context components.
- * @param events Mocked events for testing
- * @returns A test context for testing context component
- */
-export function makeTestComponentContext<Events = DefaultEvents>(
-  events: Events,
-): TestComponentContext<Events> {
-  // Types here are just for examing test results
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const effects: Array<[Observable<unknown>, (item: any) => void]> = []
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const immediateEffects: Array<[Observable<unknown>, (item: any) => void]> = []
-  const context: Mat<Events> = {
-    [matType]: 'tester',
-    events,
-    bindEffect: (observable, effect) => effects.push([observable, effect]),
-    bindImmediateEffect: (observable, effect) =>
-      immediateEffects.push([observable, effect]),
-    jsx: null, // TODO: tester JSX function
-  }
-  return { context, effects, immediateEffects }
-}
-
-/**
- * Does an element description have any binds?
- *
- * @param description Element description
- * @returns True if any dynamic binds
- */
-export function hasAnyBinds(description: ElementDescription): boolean {
-  return (
-    Boolean(description.childrenBind) ||
-    Object.keys(description.bind).length > 0 ||
-    Object.keys(description.immediateBind).length > 0 ||
-    Object.keys(description.events).length > 0 ||
-    Object.keys(description.styleBind).length > 0 ||
-    Object.keys(description.immediateStyleBind).length > 0 ||
-    Object.keys(description.classBind).length > 0 ||
-    Object.keys(description.immediateClassBind).length > 0
-  )
 }
